requirements.txt
# Core Dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5  # 支持 DeepSeek（OpenAI 兼容 API）
langchain-community>=0.0.20
pydantic>=2.5.0
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.8.0
tenacity>=8.2.3  # For retry logic

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
mypy>=1.7.0

